add runtime logging
