+++ lib/Compiler.js
@@ -28,6 +28,7 @@ const RequestShortener = require("./RequestShortener");
 const { makePathsRelative } = require("./util/identifier");
 const ConcurrentCompilationError = require("./ConcurrentCompilationError");
 const { Logger, LogType } = require("./logging/Logger");
+const logToConsole = require("./logging/logToConsole");
 
 /** @typedef {import("../declarations/WebpackOptions").Entry} Entry */
 /** @typedef {import("../declarations/WebpackOptions").WebpackOptions} WebpackOptions */
@@ -220,92 +221,9 @@ class Compiler extends Tapable {
 				}
 			}
 			if (this.hooks.log.call(name, type, args) === undefined) {
-				const labeledArgs = (prefix = "") => {
-					if (Array.isArray(args)) {
-						if (args.length > 0 && typeof args[0] === "string") {
-							return [`${prefix}[${name}] ${args[0]}`, ...args.slice(1)];
-						} else {
-							return [`${prefix}[${name}]`, ...args];
-						}
-					} else {
-						return [];
-					}
-				};
-				switch (type) {
-					case LogType.debug:
-						// ignore
-						break;
-					case LogType.log:
-						console.log(...labeledArgs());
-						break;
-					case LogType.info:
-						console.info(...labeledArgs("<i> "));
-						break;
-					case LogType.warn:
-						console.warn(...labeledArgs("<w> "));
-						break;
-					case LogType.error:
-						console.error(...labeledArgs("<e> "));
-						break;
-					case LogType.trace:
-						console.trace();
-						break;
-					case LogType.group:
-						// eslint-disable-next-line node/no-unsupported-features/node-builtins
-						if (typeof console.group === "function") {
-							// eslint-disable-next-line node/no-unsupported-features/node-builtins
-							console.group(...labeledArgs());
-						} else {
-							console.log(...labeledArgs());
-						}
-						break;
-					case LogType.groupCollapsed:
-						// eslint-disable-next-line node/no-unsupported-features/node-builtins
-						if (typeof console.groupCollapsed === "function") {
-							// eslint-disable-next-line node/no-unsupported-features/node-builtins
-							console.groupCollapsed(...labeledArgs());
-						} else {
-							console.log(...labeledArgs("<g> "));
-						}
-						break;
-					case LogType.groupEnd:
-						// eslint-disable-next-line node/no-unsupported-features/node-builtins
-						if (typeof console.groupEnd === "function") {
-							// eslint-disable-next-line node/no-unsupported-features/node-builtins
-							console.groupEnd();
-						} else {
-							console.log(...labeledArgs("</g> "));
-						}
-						break;
-					case LogType.time:
-						console.log(
-							`[${name}] ${args[0]}: ${args[1] * 1000 + args[2] / 1000000}ms`
-						);
-						break;
-					case LogType.profile:
-						// eslint-disable-next-line node/no-unsupported-features/node-builtins
-						if (typeof console.profile === "function") {
-							// eslint-disable-next-line node/no-unsupported-features/node-builtins
-							console.profile(...labeledArgs());
-						}
-						break;
-					case LogType.profileEnd:
-						// eslint-disable-next-line node/no-unsupported-features/node-builtins
-						if (typeof console.profileEnd === "function") {
-							// eslint-disable-next-line node/no-unsupported-features/node-builtins
-							console.profileEnd(...labeledArgs());
-						}
-						break;
-					case LogType.clear:
-						// eslint-disable-next-line node/no-unsupported-features/node-builtins
-						if (typeof console.clear === "function") {
-							// eslint-disable-next-line node/no-unsupported-features/node-builtins
-							console.clear();
-						}
-						break;
-					default:
-						throw new Error(`Unexpected LogType ${type}`);
-				}
+				// ignore debug logging by default
+				if (type === LogType.debug) return;
+				logToConsole(name, type, args);
 			}
 		});
 	}
