+++ lib/logging/Logger.js
@@ -31,6 +31,8 @@ const LogType = Object.freeze({
 
 exports.LogType = LogType;
 
+/** @typedef {LogType} LogTypeEnum */
+
 const LOG_SYMBOL = Symbol("webpack logger raw log method");
 const TIMERS_SYMBOL = Symbol("webpack logger times");
